decimal fixation

+++ RR_RoadWays_Services/Models/VoucherDieselDetails.cs
@@ -8,7 +8,7 @@ namespace RR_RoadWays_Services.Models
         public int Id { get; set; }
         public DateTime? Date { get; set; }
         public int? StationId { get; set; }
-        public int? Litre { get; set; }
+        public decimal? Litre { get; set; }
         public decimal? Rate { get; set; }
         public string OilAndOthers { get; set; }
         public decimal? Amount { get; set; }
